backend/pg: Stop using legacy helper/schema
The legacy SDK is a very heavy dependency that this backend was only using
tiny parts of.

We'll now instead use our new "backendbase" package, which aims to provide
a smaller set of helper functions that cover the main use-cases that the
existing backends were relying on the SDK to meet, but with considerably
less code and fewer layers of abstraction.