update test fixtures
There were two failing ACC tests with this update. One from from a newer
version of pg, which causes a different auth error message. The other is
is directly related to this PR, which was checking for a specific schema
error, which is now `invalid value for "skip_schema_creation"` when a
non-bool argument is present.