terraform: Never prune "unused" ephemeral resource nodes
I'm honestly not really sure yet how to explain _why_ ephemeral resource
nodes are getting pruned when they shouldn't; for the sake of prototyping
this is just a hard-coded special exception to just not consider them
at all in the pruneUnusedNodesTransformer.

The later ephemeralResourceCloseTransformer has its own logic for deciding
that an ephemeral resource isn't actually needed in the current graph
and pruning both their open and close nodes, so these will still get
pruned but it will happen in different circumstances and based on a later
form of the graph with more nodes and edges already present, thus
preventing some cases of ephemeral resources being pruned when they
shouldn't be.