command: "terraform apply" accepts variable values with saved plan
To support ephemeral values we need a more complicated set of rules about
what input variables can and must be set when applying a saved plan. The
command itself does not have enough information to implement those rules
itself, so we'll let them pass through and check this in the local
backend's apply phase instead.

The local backend's apply phase already has basic support for dealing with
apply-time variable values, so this just removes the blocker that was
preventing values from reaching that logic.