builtin/providers/terraform: Prepare for more ephemeral resource types
Instead of a test for whether the type name is different than the one we
expect, we'll use a switch statement. This does nothing for now, but
a future commit will add a new ephemeral resource type that's intended only
for prototyping, exploiting the fact that this particular provider can
offer ephemeral resource types without us first extending the provider
plugin protocol with that concept.